Added 2 function pathExist() && checkPath()
checkPath() >> Take 2 Parameters as input and concatenate both strings
into one string with name full path then after check that path is exist
or not.

pathExist() >> Take 1 parameter as input and check whether the path is
exist in your local computer or not will return true if it is exist and
written false if it is not exist.