Updated printlgo() function such that
Will print the status of that file with a symbol with margin of -14
strings then after the path of that file with the status, colour shown
below.

COPIED         -> file will Display log in Cyan.
UPDATED        -> file will Display log in Green.
UP TO DATE     -> file will Display log in Green.
FILE NOT FOUND -> file will Display log in Green.